Change to allow all resource for policy
The authorizer would only let the first method that called it to be
authorized, and then every call after would throw a 403 error. In order
to have the authorizer work for multiple methods, we have to change the
resource to allow all in the policy document. However, this means that
this authorizer isnt really an authorizer and more of an authenticator.