Use client secret flow for api access
This change also sets up the rotator lambda to rotate the very
credentials that it relys on. Seemed fair to make it work the same way!
Should serve as an example for other clients once it gets code reviewed
and if folks agree i didn't make any super bad mistakes with it