Refactor to allow multiple entry points
Since a deploy will currently revert client secrets i figured it would
be a good idea to trigger an immediate rotate action from codepipeline.

This could also be turned into a cfn custom resource that lives with the
project if that makes more sense. That would certainly budle the
rotation with the deploy a bunch better, but also is a bit more
cumbersome so i didn't want to run that direction till it seemed more
required