Logging refactor and cp hook expansion
I meant to make these different commits!

This should add much better logging into the lambdas. I'm playing with a
json formatter for it to see if that would make anything nicer. Probably
not worth the overhead all and all...

Also refactored the cp lambda entry to be more generic.. instead of just
being for rotating client secrets it supports an Action list and that is
one of them. I was concerned that the deploy was getting just a tad long
and didn't want to add another lambda for each possible action