Refactor for direct user grant
This refactors the auth logic to use a direct user grant and also pulls
all of the ssm stuff out of the apiproxy class. That shouldn't have been
there in the first place; it was a weird coupling.

Since i nuked the fancy json logging format in an earlier commit the
logging_setup file didn't make much sense so i killed it too