Disable dvader in upper environments
I was removing the user entirely, but I was getting really weird deploy
failures on stage/prod and think i narrowed it down to the master dvader
user getting nuked from the master realm that's imported.

I have nothing to back up that not having any users in the config causes
the problem, but circumstantially it lines up. Only difference between
stage/prod and the lower environments is rds aurora and this extra
transform step. When manually restoring the user to the config and
re-importing it resolved the deploy issue so it seems like the culprit