Adjust unhealthy threshold for eu deploys
This is absolutely not the correct fix for this.. I think the root
problem is that the ecr repo is in us-west-2 and pulling this sort of
large image over to eu-central-1 is adding enough time that the LB marks
the containers as unhealthy and pulls them just before they would be
fully initialized.

If confirmed I'll revert this and update the pipeline to publish the
container to eu-central-1. Till then though this should make deploying
to upper environments doable without manual intervention