Bakes in NR agent defaults
Disables browser monitoring because kc shouldn't be having any actual
frontend traffic that we'd need to inspect akaik.

Also turns off logging. We don't have a mechanism in place to retrieve
the logs from the container so it'll just be spending cycles/storage
writing to a file we can't retrieve. In dev (or aws) we can re-enable
with an env var in the task def