Add external navexdev network
Adds a new external docker network via dev-deploy and modifies
keycloak docker-compose to spin up the Keycloak containers on that
network. This is to support other external containers like the ones
for COI to be able to connect to keycloak without having to spin
Keycloak up inside their docker-compose files.