Only rotate confidential client secrets
KC stores a client secret even for clients that don't use them and they
were getting picked up in the rotation logic and mirrored to ssm even
though they aren't used for anything.

This should just cause a bit less ssm api traffic and i was noticing a
bit of rate limiting in the logs that could possibly be causing an issue
on occasion.