Adds start/end time to config import log search
AWS recently introduced a bug in the FilterLogEvents endpoint that the
config importer uses that causes it to not find logs if a reasonably
small time range isn't specified in the search. This now searches in the
prior hour of logs which should be more than enough since the import
ought to time out after 15 minutes..

There was also a bug in the time out logic that this fixes. Python
timestamps return as a floating number with second precision on the
left, but aws generally deals in integers with millisecond precision.
When checking if the import had timed out i forgot to convert the python
timestamp to milliseconds so the comparison wasn't very close.