Update Postgres to 11.16
AWS is deprecating postgres 10.x in early 2023 and that is a good enough
reason to get on something less ancient. This change will require out of
band migration steps in aws before it can succeed in any environment due
to limitations in cfn rds support when changing
versions/parameter-groups/instance classes.