Disable client cert auth requesting
This setting gets enabled from this bundled cli script:
  https://github.com/keycloak/keycloak-containers/blob/15.0.2/server/tools/cli/x509-truststore.cli#L13
when you specify the X509_CA_BUNDLE env var to the container.. we
started doing that on dev/lower environments awhile ago when
applications started to implement backchannel logout and for some reason
it was causing only dev to request client certs on first page visit.
This change disables that config setting; i am not positive that it
isn't needed, but suspect it is just a convienence config change because
client cert auth is the main reason folks would specify the CA Bundle
config. Worst case though that setting is only needed in lower
environments so the blast radius is pretty minimal