Update obiwandc.cer
Add empty trailing line. I think this is breaking the cert store trust
logic. The bundled script that handles that uses the
`-----BEGIN CERTIFICATE-----` line as a delimiter and I think might 
be failing because it ends up with the BEGIN line for the root CA
on the same line as the END line from obiwan:
https://github.com/keycloak/keycloak-containers/blob/15.0.2/server/tools/x509.sh#L73-L74

Still a bit of a shot in the dark though